stable 9.6: push bugfix from dev to stable build
6d49e295 fix install_from_github func
361e122e impl https://github.com/Ansuel/tch-nginx-gui/issues/945
779cf3c3 gui: move gui upgrade option as the first option
98b79a91 Revert "gui: move gui upgrade option as the first option"
db965bfe core: add check for mkdir and normalize logging file
3dc5d0c2 gui: move gui upgrade option as the first option
51f5252b }?
753371a6 update changelog issue template
36aff9fc fix syntax error move bank info to modgui modal
d0fe8fa3 remove also mobiled mappers when lib not found more shell
script optimizations
04ecc69e more infos for debug file minor shell warnings for rootdevice
fix https://github.com/Ansuel/tch-nginx-gui/issues/1034
0c0ca39d core: fix error in remove modgui_log dir
ec9e9da4 core: improve loggin of data
2bc05f59 core: move console echo to dedicated function
431e48a5 syntax error
db8730f9 force bank1 erase and reboot also on first install from bank2
check /proc/prozone is exposed before tricking failboot to avoid
warnings
5d48b3a9 fix feeds for 17.1/2 arm added macoers feeds for arm 18.x
update DGA (arm) specific package embedded ipk fix mv scripthelpers
warnings during install on some devices remove mobiled card on
unsupported devices (mobiled not installed)
a5f2b677 fix "1" echo during install
549568e0 fix perm
669b508f missing source file
bab60ea8 github: fix typo in README (#1026)
f38c4aa4 typo
2718536e fixes login for opkg repo feed fix logger_command: not found
during specific package install
0689a055 make specific_app cases more generic on cpu type
1e76fbab fix luci install on some firmwares
22a65191 rework extensions install checks to use marketing_version and
cputype
318d7819 gui: fix wireless-client-modal broken in minified source
eefa1fef move custom func to modgui modal remove transfer bank clone
bank
44f2bd31 gui: fix typo in serial modem for missing input box
7c5089c4 gui: permit to force a different modem serial
9ae6fee7 gui: improve cwmp settings in modgui modal
cc1ea837 gui: drop use of cwmp TIM latest version
0007ffce gui: Add new option to disable cwmp update
5c575439 gui: fix bad cwmp tim version
c1650eb8 gui: also export Kernel version in modgui mapper
43ea509f gui: export CPUtype in modgui mapper
fdac82da xdsl drivers: add A2pvfbH046w B2pvfbH046w
a2ce5dd0 remove hardcoded version from rootdevice
3986f23f typo
27f20328 core: disable switchover for cwmp
49281c6c core: skip bank_1 erase if it doesn't contain a valid firmware
9c38099f gui: drop use of driver_version to store xdsl firmware version
6f318fb8 gui: remove use of versioncustom latest config
2b262db7 gui: remove use of gui_version in modgui config
fa5333d0 gui: move gui version to rpc mapper
4787ef78 core: add option to disable modoverlay
317b3959 core: force otp to second bank
d9210a9e improve feeds detection
3ab577c9 core: force second bank on firmware upgrade
5e5dcd1c core: fix typo in cwmpd upgrade script
e6506fdb shellcheck avoid subshell spawning
fb149c3c core: refresh firmware upgrade script
da64dfce core: refresh cwmpd upgrade script
708bedf2 core: fix wrong preinit check
e8091fbc core: update core file on gui upgrade
80094ba2 Fixed FEC translations (#1016)

+++ decompressed/base/usr/lib/cwmpd/transfers/upgrade.sh
@@ -1,9 +1,28 @@
 #!/bin/sh
 
+. /lib/functions.sh
+
 SCRIPT_DIR=$(dirname $0)
-
 ERROR_FILE=$1
+CONFIG=cwmp_transfer
 
+
+char_encoding()
+{
+  lua <<EOF - $1
+local name = arg[1]
+local escapes = ":/?#[]@!$&'()*+;="
+local function escape_char(x)
+  if escapes:find(x, nil, true) then
+    return ("%%%02X"):format(x:byte(1))
+  else
+    return x
+  end
+end
+name =name:gsub(".", escape_char)
+io.write(name)
+EOF
+}
 
 do_upgrade()
 {
@@ -12,7 +31,10 @@ do_upgrade()
   WAIT_FOR_SWITCHOVER_FILE="/usr/lib/cwmpd/transfers/cwmp_waitforswitchover"
   rm -f $WAIT_FOR_SWITCHOVER_FILE
   if [ -x /usr/bin/sysupgrade-safe ]; then
+    # Modgui: detected acs upgrade, skip version spoof
+    # (use the real firmware version in versioncusto)
     echo 1 > /overlay/.skip_version_spoof
+
     if ! [ -z "$(grep bank_2 /proc/mtd)" ]; then
       $SCRIPT_DIR/rollback.sh record
       if [ "$(uci get cwmpd.cwmpd_config.upgrade_switchovertype)" = "1" ]; then
@@ -21,10 +43,11 @@ do_upgrade()
     else
       WAIT_FOR_SWITCHOVER_FILE="/"
     fi
+
     if [ $SWITCHOVER_TYPE = "delayed" ]; then
       /usr/bin/sysupgrade-safe -o "$url"
     else
-    /usr/bin/lua /usr/lib/cwmpd/transfers/checkOngoingServices.lua
+      /usr/bin/lua /usr/lib/cwmpd/transfers/checkOngoingServices.lua
       /usr/bin/sysupgrade-safe "$url"
     fi
 
@@ -34,8 +57,8 @@ do_upgrade()
       if [ $SWITCHOVER_TYPE = "delayed" ]; then
         touch $WAIT_FOR_SWITCHOVER_FILE
       fi
-      exit 0    
-    else           
+      exit 0
+    else
       return $UERR
     fi
   else
@@ -43,12 +66,6 @@ do_upgrade()
   fi
 }
 
-CONFIG=cwmp_transfer
-
-char_encoding()
-{
-  echo "$1" | sed -e 's/:/%3A/g' -e 's/\//%2F/g'  -e 's/?/%3F/g' -e 's/#/%23/g' -e 's/\[/%5B/g'  -e 's/\]/%5D/g'  -e 's/@/%40/g'  -e 's/!/%21/g'  -e 's/\$/%24/g'  -e 's/&/%26/g' -e 's/'\''/%27/g' -e 's/(/%28/g' -e 's/)/%29/g' -e 's/*/%2A/g' -e 's/+/%2B/g' -e 's/,/%2C/g' -e 's/;/%3B/g' -e 's/=/%3D/g'
-}
 get_url()
 {
   local URL=$1
@@ -62,7 +79,7 @@ get_url()
   fi
 
   if [ ! -z $password ]; then
-    password=$(char_encoding "$password") 
+    password=$(char_encoding "$password")
     usrinfo="$usrinfo:$password"
   fi
 
@@ -83,20 +100,6 @@ get_uci_id()
     id=
   fi
   echo $id
-}
-
-is_dual_bank()
-{
-  grep -c "\"bank_2\"" /proc/mtd >/dev/null
-}
-
-# store the expected active bank after the upgrade
-remember_bank()
-{
-  local id="$1"
-  if is_dual_bank && [ -f /proc/banktable/notbooted ] ; then
-    uci set $CONFIG.$id.bank=$(cat /proc/banktable/booted)
-  fi
 }
 
 get_started()
@@ -138,7 +141,7 @@ set_started()
     if [ ! -z $url ]; then
       uci set $CONFIG.$id.url="$url"
     fi
-    remember_bank $id
+    [ -n "$remember_bank" ] && remember_bank $id
     uci commit
     return
   fi
@@ -203,6 +206,16 @@ fi
 if [ -z $TRANSFER_URL ]; then
   echo "no URL specified"
   ubus send FaultMgmt.Event '{ "Source":"cwmpd", "EventType":"ACS provisioning", "ProbableCause":"Firmware upgrade error", "SpecificProblem":"no URL specified" }'
+
+  if [ -f  /var/state/cwmpd ]; then
+    local failure_count
+    config_load cwmpd
+    config_get failure_count cwmpd_config acs_upgrade_failures "0"
+    failure_count=$(( failure_count + 1 ))
+    sed -i "/cwmpd.cwmpd_config.acs_upgrade_failures/d" /var/state/cwmpd
+    uci -P /var/state set cwmpd.cwmpd_config.acs_upgrade_failures="${failure_count}"
+    uci -P /var/state commit cwmpd
+  fi
   exit 1
 fi
 
@@ -229,16 +242,29 @@ if [ "$TRANSFER_ACTION" = "start" ]; then
     set_started "$TRANSFER_ID" yes "$URL"
     uci show $CONFIG
 
-	ubus send cwmpd.transfer '{ "session": "begins", "type": "upgrade" }'
-    do_upgrade "$URL"
-    #if do_upgrade returns, the upgrade failed, remember that
-    if [ $? -eq 1 ]; then
-      E="1,download upgrade image failed"
+    ubus send cwmpd.transfer '{ "session": "begins", "type": "upgrade" }'
+
+    if [ "$(uci get -q mogui.var.disable_cwmp_update)" = "1" ]; then
+      # TIM IS LOVE, TIM IS LIFE. No disclusure for you :D
+      if echo $TRANSFER_URL | grep -q 'Firmware/TR069/AGThomson'; then
+        uci set modgui.var.reboot_reason_msg="TIM ACS asked for firmware update. Go to Modgui Settings to allow this."
+      else
+        echo "$TRANSFER_URL" "$TRANSFER_USERNAME" "$TRANSFER_PASSWORD" > /tmp/cwmpd_update_request
+        uci set modgui.var.reboot_reason_msg="ACS asked for firmware update from: $URL Go to Modgui Settings to allow this."
+      fi
+      exit 0
     else
-      E="1,upgrade failed (not a valid signed rbi?)"
+      do_upgrade "$URL"
+      #if do_upgrade returns, the upgrade failed, remember that
+      if [ $? -eq 1 ]; then
+        E="1,download upgrade image failed"
+      else
+        E="1,upgrade failed (not a valid signed rbi?)"
+      fi
     fi
+
     set_error "$TRANSFER_ID" "$E"
-	ubus send cwmpd.transfer '{ "session": "ends", "type": "upgrade" }'
+    ubus send cwmpd.transfer '{ "session": "ends", "type": "upgrade" }'
 
     # In case of single bank, reboot the board
     platform_is_dualbank
@@ -250,8 +276,31 @@ if [ "$TRANSFER_ACTION" = "start" ]; then
       echo b 2>/dev/null >/proc/sysrq-trigger
     fi
   else
-    #retrieve error
-    E=$(get_error "$TRANSFER_ID")
+    if [ -d $SCRIPT_DIR/target ]; then
+      id="$(get_uci_id "$TRANSFER_ID")"
+      if ! TARGET="$(uci get "$CONFIG.$id.target")"; then
+        TARGET=gateway
+      fi
+      if ! echo "$TARGET" | grep -Eqx '[A-Za-z0-9_-]+'; then
+        echo "Invalid target"
+        exit 1
+      fi
+
+      GET_ERROR_PATH="$SCRIPT_DIR/target/$TARGET/get_error"
+      if ! [ -x "$GET_ERROR_PATH" ]; then
+        echo "Missing check_error script"
+        exit 1
+      fi
+
+      STORED_ERROR="$(uci -q get "$CONFIG.$id.error")"
+      if ! E="$("$GET_ERROR_PATH" "${STORED_ERROR:-0}" "$SCRIPT_DIR" "$CONFIG.$id" "$TRANSFER_ID")"; then
+        echo "Unable to retrieve error"
+        exit 1
+      fi
+    else 
+      #retrieve error
+      E=$(get_error "$TRANSFER_ID")
+    fi
   fi
   if [ "$E" != "0" ]; then
     local msg=$(echo $E | cut -d, -f2)
@@ -259,25 +308,64 @@ if [ "$TRANSFER_ACTION" = "start" ]; then
     if [ ! -z $ERROR_FILE ]; then
       echo $msg >$ERROR_FILE
     fi
+
+    if [ -f  /var/state/cwmpd ]; then
+      local failure_count
+      config_load cwmpd
+      config_get failure_count cwmpd_config acs_upgrade_failures "0"
+      failure_count=$(( failure_count + 1 ))
+      sed -i "/cwmpd.cwmpd_config.acs_upgrade_failures/d" /var/state/cwmpd
+      uci -P /var/state set cwmpd.cwmpd_config.acs_upgrade_failures="${failure_count}"
+      uci -P /var/state commit cwmpd
+    fi
+
     ubus send FaultMgmt.Event '{ "Source":"cwmpd", "EventType":"ACS provisioning", "ProbableCause":"Firmware upgrade error", "SpecificProblem":"'"$(echo $E | cut -d, -f2)"'", "AdditionalText":"URL='"$TRANSFER_URL"'"}'
     exit 1
   fi
   ubus send FaultMgmt.Event '{ "Source":"cwmpd", "EventType":"ACS provisioning", "ProbableCause":"Firmware upgrade success", "SpecificProblem":"", "AdditionalText":"URL='"$TRANSFER_URL"'"}'
+
+  if [ -f /var/state/cwmpd ]; then
+    local upgrade_time=$(date "+%FT%TZ")
+    sed -i "/cwmpd.cwmpd_config.acs_last_upgrade_time/d" /var/state/cwmpd
+    uci -P /var/state set cwmpd.cwmpd_config.acs_last_upgrade_time="${upgrade_time}"
+    uci -P /var/state commit cwmpd
+  fi
+
   exit 0
 fi
 
 if [ "$TRANSFER_ACTION" = "cleanup" ]; then
-  set_started "$TRANSFER_ID" no
-  # Remove RAW storage information if coming from legacy SW (transfer ID is hexadecimal)
-  if [ -f /proc/banktable/legacy_upgrade/key ]; then
-    [ "$TRANSFER_ID" == $(cat /proc/banktable/legacy_upgrade/key | hexdump -v -e '/1 "%02X"') ] && echo "1" > /proc/banktable/erase_upgrade_info
+  if [ -d $SCRIPT_DIR/target ]; then
+    id=$(get_uci_id "$TRANSFER_ID")
+    if ! TARGET="$(uci get "$CONFIG.$id.target")"; then
+      TARGET=gateway
+    fi
+    if ! echo "$TARGET" | grep -Eqx '[A-Za-z0-9_-]+'; then
+      echo "Invalid target"
+      exit 1
+    fi
+
+    CLEANUP_PATH="$SCRIPT_DIR/target/$TARGET/cleanup"
+    if ! [ -x "$CLEANUP_PATH" ]; then
+      echo "Missing cleanup script"
+      exit 1
+    fi
+
+    set_started "$TRANSFER_ID" no
+    exec "$CLEANUP_PATH" "$SCRIPT_DIR" "$CONFIG.$id" "$TRANSFER_ID"
+  else
+    set_started "$TRANSFER_ID" no
+    # Remove RAW storage information if coming from legacy SW (transfer ID is hexadecimal)
+    if [ -f /proc/banktable/legacy_upgrade/key ]; then
+      [ "$TRANSFER_ID" == $(cat /proc/banktable/legacy_upgrade/key | hexdump -v -e '/1 "%02X"') ] && echo "1" > /proc/banktable/erase_upgrade_info
+    fi
+    # Upgrade finished succesfully, remove database and transfer information from passive bank (for dual bank platform)
+    if is_dual_bank; then
+      [ -f /overlay/$(cat /proc/banktable/notbooted 2>/dev/null)/etc/cwmpd.db ] && rm /overlay/$(cat /proc/banktable/notbooted)/etc/cwmpd.db
+      [ -f /overlay/$(cat /proc/banktable/notbooted 2>/dev/null)/etc/config/cwmp_transfer ] && rm /overlay/$(cat /proc/banktable/notbooted)/etc/config/cwmp_transfer
+    fi
+    exit 0
   fi
-  # Upgrade finished succesfully, remove database and transfer information from passive bank (for dual bank platform)
-  if is_dual_bank; then
-    [ -f /overlay/$(cat /proc/banktable/notbooted 2>/dev/null)/etc/cwmpd.db ] && rm /overlay/$(cat /proc/banktable/notbooted)/etc/cwmpd.db
-    [ -f /overlay/$(cat /proc/banktable/notbooted 2>/dev/null)/etc/config/cwmp_transfer ] && rm /overlay/$(cat /proc/banktable/notbooted)/etc/config/cwmp_transfer
-  fi
-  exit 0
 fi
 
 echo "Unknown transfer action: $TRANSFER_ACTION"
